Move resetStorage() function to its own file that's only imported from the browser extension context.
